Add test cases to confirm @Id population issue
After further investigation, a strange behavior
related to the @Id population was found.
Whenever the @Id field is a val, and the model
has more than fifteen attributes (not counting
the @Id field), the CRUDRepository save and
saveAll methods' returned instance doesn't have
the @Id field populated.
Despite the collection document having that field
populated, there is an inconsistency between the
returned instance and the database entry.
With that in mind, this commit adds a test class
that confirms the described behavior.